requirements.txt
requests==2.32.3
python-dotenv==1.1.0 
google-cloud-texttospeech==2.25.1
google-generativeai==0.8.4
googlemaps==4.10.0
pandas==2.2.3
openpyxl==3.1.2
